+++ NEWS.md
@@ -1,5 +1,7 @@
 # openxlsx (development version)
 
+* Added optional argument to `loadWorkbook` to decide if empty/blank cells should be converted to NA_character_ (the default) or left blank as is
+
 * Add type blanks and not blanks to conditional formatting ([#311](https://github.com/ycphs/openxlsx/pull/311))
 
 # openxlsx 4.2.5